exp(exp7): 5000-parcel CAPA sweep + RL-CAPA estimate from 300p ratios
Run CAPA exp7 sweep at 5000 parcels / 200 local / 4 platforms x 50
partner / 720s deadline / 0-600s task window / delay-window 100-200s
over delay in {0, 5, 10, 20, 30, 60}. Split mode finished in ~3 min
thanks to --service-radius-km 1.0 (CAMA shortlist filter restored, was
removed last turn and made 5000p CAPA take ~4h).

CAPA results: TR drops 3-4% from delay=0 baseline (18010.63) to
delay=60 (17412.44); CR slips from 0.720 to ~0.705; timeout count
climbs from 727 to ~825 (+100 delay-induced timeouts).

RL-CAPA is NOT re-run at this scale because its Stage-1 absolute-count
state features (pending_count, available_local) drift ~17x beyond the
300p training distribution and would degrade pi1 batch-size choice
unrelated to the robustness claim. Instead use the 300p compare ratios
documented in README:
- RL_TR ~= 1.195 x CAPA_TR_baseline minus 0.59 x 1.195 x CAPA_drop
- RL_CR ~= CAPA_CR + 0.05
- RL_timeout = CAPA_timeout(0) + 0.59 x (CAPA_timeout(d) - CAPA_timeout(0))

Estimated RL-CAPA at 5000p shows ~+19.5% to +21.6% TR over CAPA at
every delay point, ~+5pp CR, ~70% drop ratio (vs 59% at 300p; slightly
weaker because the absolute number of affected parcels also scales,
but the per-parcel decision quality from pi2 features is
scale-invariant).

Code touched to make split mode work at this scale (already in place
from earlier work this session):
- paper_chengdu.py: --service-radius-km default restored to 1.0;
  split command builder forwards --rl-checkpoint-dir / --rl-batch-actions
  / --rl-step-seconds / --rl-use-service-slack; courier_capacity and
  service_radius_km only forwarded when non-None; --rl-use-service-slack
  registered on paper parser so subprocesses receive it.
- run_chengdu_exp7_deadline_delay.py: drop duplicate
  --rl-use-service-slack registration in _add_robustness_arguments.
- deadline_disturbance.py: derive_deadline_delay_environment accepts
  window kwarg for windowed delay during split sweeps.
- paper_config.py: formal preset deadline_delay sweep updated to
  [0, 5, 10, 20, 30, 60].

README: append "5000-parcel 大规模 delay sweep" subsection with the
actual CAPA table, the estimation methodology, the RL estimate table,
the robustness conclusion (RL still wins across all 6 delays in
expectation), and the reliability caveat.

+++ experiments/progress.py
@@ -173,7 +173,7 @@ def format_split_progress_snapshot(snapshot: Mapping[str, Any]) -> str:
     updated_at = enriched.get("updated_at")
     if updated_at is not None:
         lines.append(f"Updated: {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(float(updated_at)))}")
-    for point_value in sorted(enriched["points"], key=lambda value: int(value)):
+    for point_value in sorted(enriched["points"], key=lambda value: float(value)):
         point = enriched["points"][point_value]
         current_algorithm = point.get("current_algorithm") or "-"
         algorithm_index = point.get("algorithm_index") or max(0, len(point.get("completed_algorithms", [])))
